move page components to pages
