+++ web/pages/2018/index.tsx
@@ -1,2 +1,2 @@
-import TopPage from '../../containers/Top';
+import TopPage from '..';
 export default TopPage;
